svn:revision:42353
svn:branch:cmake
svn:account:starseeker

+++ src/util/bw-a.c
@@ -117,7 +117,7 @@ main(int argc, char **argv)
 
     /* autosize the input? */
     if (fileinput && autosize) {
-	unsigned long int w, h;
+	size_t w, h;
 	if (fb_common_file_size(&w, &h, file_name, 1)) {
 	    file_width = (long)w;
 	} else {
